requirements.txt
# Network Anomaly Detection System - Requirements
# Core Dependencies
flask>=3.0.0
scapy>=2.6.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
psutil>=5.9.0

# Database
sqlalchemy>=2.0.0

# Web Interface
jinja2>=3.1.0
wtforms>=3.1.0

# Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Network and System
netifaces>=0.11.0

# Machine Learning
joblib>=1.3.0

# Configuration and Logging
pyyaml>=6.0.0

# Windows Service
pywin32>=306

# Security
cryptography>=41.0.0
bcrypt>=4.1.0

# API and Communication
requests>=2.31.0

# Data Processing
python-dateutil>=2.8.0
pytz>=2023.3
